+++ src/config/cli.rs
@@ -1,4 +1,4 @@
-use clap::{Parser, Subcommand, ValueEnum};
+use clap::{Parser, Subcommand, ValueEnum, ValueHint};
 use std::path::PathBuf;
 
 #[derive(Parser, Debug)]
@@ -101,7 +101,15 @@ pub enum Command {
     /// Clear the current queue
     Clear,
     /// Add a song to the current queue. Relative to music database root. '/' to add all files to the queue
-    Add { file: String },
+    Add {
+        /// Files to add to MPD's queue
+        #[arg(value_hint = ValueHint::AnyPath)]
+        files: Vec<PathBuf>,
+        /// Rmpc checks whether MPD supports the added external file's extension and skips it if it does not.
+        /// This option disables this behaviour and rmpc will try to add all the files
+        #[arg(long = "skip-ext-check", default_value = "false")]
+        skip_ext_check: bool,
+    },
     /// Add a song from youtube to the current queue.
     AddYt { url: String },
     /// List MPD outputs